moving dem test to right folder
